requirements.txt
qrcode[pil]==8.0
Pillow>=10.4.0
requests>=2.32.0